Fix some file naming and import bugs:
While doing some refactoring in React, I renamed
a lot of component filenames to CamelCase style as
this seems to be a common pattern in React.

However, git doesn't seem to pickup file renames
that don't actually change the letters used (i.e.
meeting.jsx -> Meeting.jsx).

Building locally, there were no issues. But the heroku
buildpack for Ruby wasn't able to find the imports because
of the naming issue. The solution was to run the command
`git mv file File` which made git track the filename change.