Sign up/Sign in tweaks and fixes:
- Explicitly redirect to the login page after sign up. Devise was
automatically redirecting to the root_path which then redirected to
the login page because root_path requires an authenticated user.

This didn't matter so much before, but with the :confirmable module
for devise activated, it meant that the user wasn't seeing the message
telling them to check their email for their activation link. Devise
was redirecting them and telling them they need to sign in to continue.
If the user didn't know any better, they'd have no way of knowing they
had to check their email.

- Change the sign up/sign in headers to h4 instead of h2 to better match
the rest of the application.

- Style the "Didn't receive confirmation instructions page."