Fix uploads being removed from agendum details on agendum update.
I originally kept agendum uploads nested under agendums to keep
the front-end code a little more simpler. However, this led to
uploads suddenly disappearing from the agendum details whenever
you updated an agendum (e.g. changed the title). This is because
the server was returning the updated agendum, sans uploads, which
was then being set in the client-side.

The solution I decided on was to separate uploads from the agendum
state in the client, like I did with agendum notes, as this will
allow for more flexibility and doesn't put the extra burden on
the server.