Automatically open the agendum details section in the sidebar when an agendum is selected.
Currently I'm doing something I probably shouldn't be, which is
to call a JavaScript function within rails assets instead of
defined within React/the app/javascript folder.

I'm doing this for now to make use of the materializecss framework
in use and plan to remove this smell to make the meeting page
less or non-dependent on the materializecss framework.