Finished routes refactoring:
- Use shallow routes where necessary to avoid long URLs such as
`/meetings/:meeting_id/agenda/:agenda_id/...` which are ugly and long
to type in the JavaScript files.

- Move `/uploads` into a scope under `/agenda` to logically group
the URLs.

- Move `/uploads/:id/download` to the root-level as it's not
dependent on any other model.

- Removed some comments from controllers while refactoring as
I added a lot of comments when I was first getting started with
this project. But I feel now like the comments I added weren't
adding any value as all that information could be found with
`rails routes` or by simply reading the code.