Fix a duplicate query being run on `MeetingsController.show`:
`MeetingsController.show` loads a meeting and preloads all the
data that will be needed, which generated a query with quite
a few joins with agendums and agenum notes ordered by create
date.

Howewever, using `.order` was causing a database query to be
triggered that only returned the fields being ordered on.
This caused a second query to be executed not long after,
when the view tried to render data which wasn't available.

Instead of rewriting the ActiveRecord query into a more
complex piece of code, I've opted to sort the results that
need sorting on the client side. This doesn't seem to add
any noticable overhead and it'll reduce cognitive load on
the server side.