Move creation of meetings out of the React component.
Having the React component handle both creation and editing of
meetings is causing issues with some JavaScript not being
applied to some components because they aren't in the DOM until
the meeting is actually created.

Moving non-existing meetings out of the React component will
also allow slimming down that component.