Fix "Invalid Date" being displayed in non-Chrome browsers.
It seems that the acceptable date format is dependable on the
browser so momentjs was able to correctly ready and convert
dates in Chrome but not other browsers. This resulted in
"Invalid Date" being displayed anywhere I was converting to
local time with momentjs.

For reference: https://stackoverflow.com/a/4310986