Fix multiple JavaScript functions not working in Safari.
Things such as adding attendees and agendums weren't working in
Safari. It seems that I was missing a closing brace in a query
selector, which Safari was strict on, causing Ajax requests to
fail because the CSRF token wasn't beening submitted.

Although the following link is for jQuery specifically, it lead
me in the right direction: https://stackoverflow.com/a/43209100