Create meetings from Google Calendar events:
Allow users to create a meeting and add attendees for the meeting
with the single click of a button.

Note that currently the calendar event timezones aren't converted
to the user's timezone or even UTC. Doing so would unfortunately
require an extra API call to pull the calendar's timezone as that
information isn't always stored in the event itself.

For now I'd like to use the feature for awhile before deciding
how to go about handling this issue.