Cosmetic updates for the recent and upcoming meetings section.
- Add 'no recent/upcoming meetings' message to the recent and
upcoming meetings panels intead of just not displaying anything.

- Remove the recent/upcoming meetings panels when not on page 1.

If you're on a page other than 1 it's most likely that there will
be no meetings that fall into either of those categories as the
filtering is done on the meetings that have already been loaded
for that page.

So having those sections may confuse users as either not all their
recent/upcoming meetings will show up there or nothing will show
up. Always showing the full upcoming/recent meetings would require
separate queries for each resulting in 3 queries per page just for
the meeting lists which may become a bottleneck in the future for
something that may not matter too much to the average user.

Therefore, I've decided to only show those sections on the first
page for now.