Remove hack for more "material-like" tables.
The table design in materialize css isn't up-to-date with the
recent table design from the spec. However, the collection design
looks similar, so I was making a table also a collection to get
the design benefits.

However, it later hit me that I can make the tables look more
material-like with just a few CSS changes. Since extra CSS was
required for the collection hack anyways and not using collections
is more correct and easy to understand, I've made tables look more
material-like without collections.