Convert <TimePicker /> to track its own internal state
There were issues being caused by it relying completely on the props
being passed in. Mainly, if you entered something that didn't format
correctly out of morment, it would come up with random values.

For example, if you tried to write '03' for '03:00 PM' it might
spit out '12' changing it to '12:00' which was confusing and
often required pressing the numbers you want a few times to
actually get the numbers you wanted to show.