Fix error on signup page visits
Users were getting errors when visiting the signup page. It seems
that clicking around in the "guest area" (e.g. forgot my password
page) would cause the layout to be set to the application layout
which would then error because it was trying to load the user record
and use its email address for the sidebar.